+++ src/App.tsx
@@ -1,13 +1,8 @@
 import './App.css';
 import React from 'react';
-import NameInput from './components/NameInput';
 
 function App() {
-  return (
-    <>
-      <NameInput />
-    </>
-  );
+  return <></>;
 }
 
 export default App;
